backend/requirements.txt
cryptography
sqlalchemy
python-multipart
faiss-cpu